pyproject.toml
[project]
name = "kaggle-luxai-s3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "blosc==1.11.2",
    "lmdb==1.6.2",
    "numpy>=2.1.2",
    "opencv-python==4.11.0.86",
    "pandas>=2.2.3",
    "pytorch-lightning>=2.5.0.post0",
    "pyyaml>=6.0.2",
    "torch==2.6.0+cu118",
    "transformers==4.48.2",
    "wandb>=0.19.8",
]

[tool.uv]
extra-index-url = ["https://download.pytorch.org/whl/cu118"]
